requirements.txt
cffi==1.15.1
cryptography==38.0.1
flake8==5.0.4
mccabe==0.7.0
oracledb==1.1.1
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
python-dotenv==0.21.0
